# docs/api.yaml
openapi: 3.0.3
info:
  title: Agent Pro Ghana API
  description: |
    REST API for Agent Pro Ghana — One App. Every Mobile Money Business.

    ## Authentication
    All protected endpoints require a Bearer JWT token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```

    Access tokens expire in **15 minutes**. Use the `/auth/refresh` endpoint with
    your refresh token to obtain a new access token without re-logging in.

    ## Roles
    | Role | Description |
    |------|-------------|
    | `superuser` | Platform administrator — full access |
    | `business_owner` | Company owner — manages branches, agents, reports |
    | `manager` | Branch manager — monitors agents and float |
    | `agent` | Field agent — processes MoMo transactions |
    | `auditor` | Read-only access to company data |

    ## Security Notice
    This API **never** requests, stores, or transmits a MoMo PIN.
    PIN entry happens exclusively on the official network USSD overlay.

  version: 2.0.0
  contact:
    name: Agent Pro Ghana Support
    email: dev@agentproghana.com
    url: https://agentproghana.com
servers:
  - url: https://api.agentproghana.com/v1
    description: Production
  - url: http://localhost:3000/v1
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - BearerAuth: []
paths:
...
  /admin/config/{key}:
    patch:
      tags: [Admin]
      summary: Update a system configuration value (superuser only)
      parameters:
        - in: path
          name: key
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value]
              properties:
                value:
                  type: string
      responses:
        '200':
          description: Config updated
...
  /admin/ussd-templates/{id}:
    patch:
      tags: [Admin]
      summary: Update a USSD template without requiring app update (superuser only)
      description: |
        Templates are dialed as a single combined USSD string per
        transaction — Android's public USSD API cannot reply to an
        already-open interactive session, so multi-step menu navigation
        is not possible without AccessibilityService (which this app
        does not use). See migration 002_ussd_single_dial_redesign.sql.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ussd_string_pattern:
                  type: string
                  example: '*170*1*2*{customer_phone}*{amount}#'
                  description: |
                    Single USSD string with {placeholder} tokens, dialed as
                    ONE request. Must NEVER contain a {pin} placeholder —
                    rejected with 422 if it does.
                placeholder_fields:
                  type: array
                  items:
                    type: string
                  example: ['customer_phone', 'amount']
                  description: Every {placeholder} used in ussd_string_pattern must be declared here.
                pin_prompt_strings:
                  type: array
                  items:
                    type: string
                  description: |
                    Substrings that indicate the network is showing a PIN
                    prompt. Cannot be empty — rejected with 422 if it is,
                    since without it the app cannot recognize when to pause.
                success_strings:
                  type: array
                  items:
                    type: string
                failure_strings:
                  type: array
                  items:
                    type: string
                timeout_seconds:
                  type: integer
                is_active:
                  type: boolean
      responses:
        '200':
          description: Template updated
        '422':
          description: |
            Validation failed — e.g. a {pin} placeholder was used,
            pin_prompt_strings was empty, or a placeholder in the pattern
            wasn't declared in placeholder_fields.
        '404':
          description: Template not found
  /admin/audit-logs:
    get:
      tags: [Admin]
      summary: Query audit logs (superuser only)
      parameters:
        - in: query
          name: action
          schema:
            type: string
        - in: query
          name: company_id
          schema:
            type: string
            format: uuid
        - in: query
          name: user_id
          schema:
            type: string
            format: uuid
        - in: query
          name: from_date
          schema:
            type: string
            format: date-time
        - in: query
          name: to_date
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Audit log entries
...
  /admin/ads/{ad_id}/moderate:
    patch:
      tags: [Admin]
      summary: Approve, reject, or publish an advertisement (superuser only)
      parameters:
        - in: path
          name: ad_id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action]
              properties:
                action:
                  type: string
                  enum: [approve_review, reject, publish]
                rejection_reason:
                  type: string
      responses:
        '200':
          description: Moderation action applied